Fix bugs for x64 platforms (std::string::npos has type size_t, not unsigned int). Add extra info in error messages. Silence a warning.
git-svn-id: file:///home/me/wd-local/bitKatalog/bitKatalog.svn/trunk@63 735b4f03-7022-4c9e-973d-5ed9b5edb13f

+++ src/fs.cpp
@@ -105,7 +105,7 @@ getFileListInDir(std::string path)
 std::string
 getLastComponentOfPath(std::string path)
 {
-    unsigned int pos;
+    size_t pos;
     pos=path.find_last_of('/');
     if (pos==std::string::npos)
         return path;
